requirements.txt
openai>=1.12.0
streamlit>=1.30.0
langchain>=0.1.10
langchain-community>=0.0.20
pyyaml>=6.0
networkx>=3.2
plotly>=5.20.0

